[clangd] allow clients to control the compilation database by passing in compilationDatabaseChanges in the 'workspace/didChangeConfiguration' request
This commit allows clangd to use an in-memory compilation database that's
controlled from the LSP client (-compile_args_from=lsp). It extends the
'workspace/didChangeConfiguration' request to allow the client to pass in a
compilation database subset that needs to be updated in the workspace.

Differential Revision: https://reviews.llvm.org/D49758


git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@338597 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 1747bfa3da8acf857313412a6f1f59cf5860babd)